Add IgnoreCopDirectives to Infl/SoftLineLength
When wrangling an old code-base it is sometimes easiest to add the check
and put

    # rubocop:disable Infl/SoftLineLength

directives at the top of "bad" files. This allows `rubocop` to check the
project OK and we can deal with the existing long lines at our own pace.

Sometimes the addition of the directive's increasing the line count of
the file by 1 makes the number of long lines sneak in under the
percentage limit, causing `rubocop` to grumble about unnecessary
disablings.

This change allows us to ignore the directive lines from the line count.

The check is naive, but I beleive "good enough".